Correct types for Size property.

+++ src/dbus_api.rs
@@ -380,7 +380,7 @@ fn create_dbus_blockdev<'a>(mut dbus_context: DbusContext) -> dbus::Path<'a> {
     let object_path = f.object_path(object_name, dbus_context.clone())
         .introspectable()
         .add(f.interface(STRATIS_DEV_BASE_INTERFACE, ())
-            .add_p(f.property::<i32, _>("Size", ())));
+            .add_p(f.property::<u64, _>("Size", ())));
 
     let path = object_path.get_name().to_owned();
     dbus_context.action_list.borrow_mut().push(DeferredAction::Add(object_path));
@@ -452,7 +452,7 @@ fn create_dbus_cachedev<'a>(mut dbus_context: DbusContext) -> dbus::Path<'a> {
     let object_path = f.object_path(object_name, dbus_context.clone())
         .introspectable()
         .add(f.interface(STRATIS_CACHE_BASE_INTERFACE, ())
-            .add_p(f.property::<i32, _>("Size", ())));
+            .add_p(f.property::<u64, _>("Size", ())));
 
     let path = object_path.get_name().to_owned();
     dbus_context.action_list.borrow_mut().push(DeferredAction::Add(object_path));
